+++ vector.cpp
@@ -1,43 +1,43 @@
 #include "vector.h"
 
-float Vec::AngleX() {
-    float a = atan2(x[1],x[0]) / 3.141593;
+double Vec::AngleX() {
+    double a = atan2(x[1],x[0]) / 3.141593;
     if (a < 0.0)
         a += 2.0;
     return a;
 }
 
-float Vec::AngleY() {
+double Vec::AngleY() {
     Vec z(0.0,0.0,1.0);
-    float a = asin(fabs(dot(z))/(magnitude())) / 3.141593 + 0.5;
+    double a = asin(fabs(dot(z))/(magnitude())) / 3.141593 + 0.5;
     if (x[2] < 0.0)
         a = 1.0 - a;
     return a;
 }
 
-float& Vec::operator[] (int i) { return x[i]; }
+double& Vec::operator[] (int i) { return x[i]; }
 Vec::Vec() {
     x[0] = x[1] = x[2] = 0.0;
 }
 
-Vec::Vec(float a, float b, float c) {
+Vec::Vec(double a, double b, double c) {
     x[0] = a;
     x[1] = b;
     x[2] = c;
 }
 
-void Vec::set(float a, float b, float c) {
+void Vec::set(double a, double b, double c) {
     x[0] = a;
     x[1] = b;
     x[2] = c;
 }
 
-float Vec::magnitude() {
+double Vec::magnitude() {
     return sqrt(x[0] * x[0] + x[1] * x[1] + x[2] * x[2]);
 }
 
 void Vec::normalize() {
-    float m = magnitude();
+    double m = magnitude();
     x[0] /= m;
     x[1] /= m;
     x[2] /= m;
@@ -45,7 +45,7 @@ void Vec::normalize() {
 
 Vec Vec::GetNormalized() {
     Vec r = *this;
-    float m = r.magnitude();
+    double m = r.magnitude();
     r[0] /= m;
     r[1] /= m;
     r[2] /= m;
@@ -60,7 +60,7 @@ Vec Vec::cross(Vec v) {
     return c;
 }
 
-Vec Vec::blend(Vec v, float f) {
+Vec Vec::blend(Vec v, double f) {
     Vec c;
     for (int i = 0; i < 3; i++) {
         c.x[i] = pow(pow(x[i],f)+pow(v.x[i],f),(1.0/f));
@@ -68,8 +68,8 @@ Vec Vec::blend(Vec v, float f) {
     return c;// / 2.0;
 }
 
-float Vec::dot(Vec v) {
-    float prod = x[0] * v.x[0] + x[1] * v.x[1] + x[2] * v.x[2];
+double Vec::dot(Vec v) {
+    double prod = x[0] * v.x[0] + x[1] * v.x[1] + x[2] * v.x[2];
     return prod;
 }
 
@@ -101,21 +101,21 @@ Vec Vec::operator-(const Vec &v) {
     c.x[2] = x[2] - v.x[2];
     return c;
 }
-Vec Vec::operator*(const float &f) {
+Vec Vec::operator*(const double &f) {
     Vec c;
     c.x[0] = x[0] * f;
     c.x[1] = x[1] * f;
     c.x[2] = x[2] * f;
     return c;
 }
-Vec Vec::operator/(const float &f) {
+Vec Vec::operator/(const double &f) {
     Vec c;
     c.x[0] = x[0] / f;
     c.x[1] = x[1] / f;
     c.x[2] = x[2] / f;
     return c;
 }
-Vec Vec::operator*=(const float &f) {
+Vec Vec::operator*=(const double &f) {
     x[0] = x[0] * f;
     x[1] = x[1] * f;
     x[2] = x[2] * f;
